requirements.txt
Flask
Flask-Cors
google-generativeai
Gunicorn
